+++ src/pages/Home.js
@@ -6,21 +6,7 @@ import CarCard from '../components/CarCard';
 
 const Home = () => {
   const cars = useSelector((state) => state.cars.cars);
-  const userPresent = useSelector((state) => state.user.id);
-  if (!userPresent) {
-    return (
-      <div className="container mt-5">
-        <div className="row justify-content-center">
-          <div className="col-md-6">
-            <div className="alert alert-danger" role="alert">
-              You need to log-in or sign-in to continue
-            </div>
-            <Link className="btn" to="/session#login">Login</Link>
-          </div>
-        </div>
-      </div>
-    );
-  }
+
   return (
     <>
       <h1>Available Cars</h1>
